+++ backend/canisters/user/c2c_client/src/lib.rs
@@ -8,7 +8,7 @@ generate_c2c_call!(c2c_edit_message);
 generate_c2c_call!(c2c_grant_super_admin);
 generate_c2c_call!(c2c_mark_read_v2);
 generate_c2c_call!(c2c_notify_group_deleted);
-generate_c2c_call!(c2c_notify_user_events);
+generate_c2c_call!(c2c_notify_events);
 generate_c2c_call!(c2c_remove_from_group);
 generate_c2c_call!(c2c_revoke_super_admin);
 generate_c2c_call!(c2c_send_messages);