+++ frontend/app/src/components/home/proposals/ProposalContent.svelte
@@ -18,7 +18,7 @@
     import ModalContent from "../../ModalContent.svelte";
     import { proposalVotes } from "../../../stores/proposalVotes";
     import { createEventDispatcher } from "svelte";
-    import type { OpenChat } from "openchat-client";
+    import { OpenChat, cryptoLookup, tokenByGovernanceCanisterLookup } from "openchat-client";
     import ProposalVoteButton from "./ProposalVoteButton.svelte";
     import ProposalVotingProgress from "./ProposalVotingProgress.svelte";
     import ChevronDown from "svelte-material-icons/ChevronDown.svelte";
@@ -38,12 +38,13 @@
     const user = client.user;
 
     const dashboardUrl = "https://dashboard.internetcomputer.org";
-    const nnsDappUrl = "https://nns.ic0.app";
 
     let summaryExpanded = false;
     let showNeuronInfo = false;
     let showPayload = false;
-
+    
+    $: token = tokenByGovernanceCanisterLookup[content.governanceCanisterId];
+    $: rootCanister = cryptoLookup[token].rootCanister;
     $: proposalTopicsStore = client.proposalTopicsStore;
     $: isNns = content.proposal.kind === "nns";
     $: voteStatus =
@@ -59,12 +60,10 @@
         proposal.status == ProposalDecisionStatus.Unspecified;
     $: proposalUrl = isNns
         ? `${dashboardUrl}/proposal/${proposal.id}`
-        : // TODO FIX THESE SNS LINKS WHEN THEY ARE AVAILABLE
-          `${nnsDappUrl}/sns/${content.governanceCanisterId}/proposal/${proposal.id}`;
+        : `${dashboardUrl}/sns/${rootCanister}/proposal/${proposal.id}`;
     $: proposerUrl = isNns
         ? `${dashboardUrl}/neuron/${proposal.proposer}`
-        : // TODO FIX THESE SNS LINKS WHEN THEY ARE AVAILABLE
-          `${nnsDappUrl}/sns/${content.governanceCanisterId}/neuron/${proposal.proposer}`;
+        : `${dashboardUrl}/sns/${rootCanister}/neuron/${proposal.proposer}`;
     $: adoptPercent = round2((100 * proposal.tally.yes) / proposal.tally.total);
     $: rejectPercent = round2((100 * proposal.tally.no) / proposal.tally.total);
     $: votingEnded = proposal.deadline <= $now;
@@ -242,19 +241,11 @@
         </ProposalProgressLayout>
     </div>
     <div class="more" class:rtl={$rtlStore}>
-        {#if isNns}
-            <a href={proposalUrl} rel="noreferrer" target="_blank">{proposal.id}</a>
-        {:else}
-            {proposal.id}
-        {/if}
+        <a href={proposalUrl} rel="noreferrer" target="_blank">{proposal.id}</a>
         <div class="subtitle">
             {typeValue} |
             {$_("proposal.proposedBy")}
-            {#if isNns}
-                <a target="_blank" rel="noreferrer" href={proposerUrl}>{truncatedProposerId()}</a>
-            {:else}
-                {truncatedProposerId()}
-            {/if}
+            <a target="_blank" rel="noreferrer" href={proposerUrl}>{truncatedProposerId()}</a>
         </div>
     </div>
 {/if}