+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -8,7 +8,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Event for setting user upgrade concurrency ([#3268](https://github.com/open-ic/open-chat/pull/3268))
-  
+- Added `OpenChatBotMessage` event type ([#3274](https://github.com/open-ic/open-chat/pull/3274))
+
+### Changed
+
+- Switch to using the new `c2c_notify_events` endpoint ([#3283](https://github.com/open-ic/open-chat/pull/3283))
+
 ### Removed
 
 - Removed code only needed for previous upgrade ([#3248](https://github.com/open-ic/open-chat/pull/3248))